Moved backend propagation code to separate file
